+++ http_request.go
@@ -10,8 +10,7 @@ import (
 )
 
 //Do is used to execute an API Request.
-func Do(req *http.Request) (string, error) {
-
+func Do(req *http.Request) (*CBResponse, error) {
 
 	var client *http.Client
 	if httpClient != nil {
@@ -22,18 +21,29 @@ func Do(req *http.Request) (string, error) {
 
 	response, err := client.Do(req)
 	if err != nil {
-		return "", err
+		return nil, err
 	}
 
 	defer response.Body.Close()
 	resBody, err := ioutil.ReadAll(response.Body)
+
+	cbResponse := &CBResponse{
+		ResponseMeta: ResponseMeta{
+			Headers:    response.Header,
+			Status:     response.Status,
+			StatusCode: response.StatusCode,
+		},
+		Body: resBody,
+	}
+
 	if err != nil {
-		return "", err
+		return cbResponse, err
 	}
 	if response.StatusCode < 200 || response.StatusCode > 299 {
-		return "", ErrorHandling(resBody)
+		return cbResponse, ErrorHandling(resBody)
 	}
-	return string(resBody), nil
+
+	return cbResponse, nil
 }
 func ErrorHandling(resBody []byte) error {
 	cbErr := &Error{}